refactored the project structure to revolve around entities and not functions.
