--- RoyalStore/src/main/java/com/royal/models/products/enumerations/Connectivity.java
@@ -1,5 +1,0 @@
-package com.royal.models.products.enumerations;
-
-public enum Connectivity {
-    FourGLTE, FiveG, WiFi, Bluetooth, NFC
-}
