--- RoyalStore/src/main/java/com/royal/models/products/enumerations/DesktopBrand.java
@@ -1,5 +1,0 @@
-package com.royal.models.products.enumerations;
-
-public enum DesktopBrand {
-    Apple, Dell, Microsoft, Asus, Lenovo, Acer, Intel, IBM, Gateway, Toshiba, HP, Fujitsu
-}
