--- RoyalStore/src/main/java/com/royal/models/products/enumerations/DesktopOS.java
@@ -1,5 +1,0 @@
-package com.royal.models.products.enumerations;
-
-public enum DesktopOS {
-    Linux, Windows, MacOS
-}
