--- RoyalStore/src/main/java/com/royal/models/products/enumerations/MobileBrand.java
@@ -1,5 +1,0 @@
-package com.royal.models.products.enumerations;
-
-public enum MobileBrand {
-    Apple, Samsung, Huawei, Xiaomi, Nokia, Motorola, Sony, Meizu, Vivo, Honor
-}
