--- RoyalStore/src/main/java/com/royal/models/products/enumerations/Processor.java
@@ -1,7 +1,0 @@
-package com.royal.models.products.enumerations;
-
-public enum Processor {
-    AppleSilicon,
-    IntelCore3, IntelCore5, IntelCore7, IntelCore9,
-    AMDRyzen3, AMDRyzen5, AMDRyzen7, AMDRyzen9
-}
