--- RoyalStore/src/main/java/com/royal/models/users/AuthenticatedUserDetails.java
@@ -1,16 +1,0 @@
-package com.royal.models.users;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.ArrayList;
-
-@Getter
-@Setter
-public class AuthenticatedUserDetails {
-    private String email;
-    private String password;
-    private boolean rememberMe;
-    private byte[] profilePicture;
-    private ArrayList<String> roles;
-}
