--- RoyalStore/src/main/java/com/royal/models/users/LoginUserCredentials.java
@@ -1,12 +1,0 @@
-package com.royal.models.users;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginUserCredentials {
-    private String email;
-    private String password;
-    private boolean rememberMe;
-}
