--- RoyalStore/src/main/java/com/royal/models/users/PublicUserDetails.java
@@ -1,17 +1,0 @@
-package com.royal.models.users;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.ArrayList;
-
-@Getter
-@Setter
-public class PublicUserDetails {
-    private String id;
-    private String token;
-    private String email;
-    private String username;
-    private byte[] profilePicture;
-    private ArrayList<String> roles;
-}
