+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/Connectivity.java
@@ -1,0 +1,5 @@
+package com.royal.products.domain.enumerations;
+
+public enum Connectivity {
+    FourGLTE, FiveG, WiFi, Bluetooth, NFC
+}
