+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/DesktopBrand.java
@@ -1,0 +1,5 @@
+package com.royal.products.domain.enumerations;
+
+public enum DesktopBrand {
+    Apple, Dell, Microsoft, Asus, Lenovo, Acer, Intel, IBM, Gateway, Toshiba, HP, Fujitsu
+}
