+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/MobileBrand.java
@@ -1,0 +1,5 @@
+package com.royal.products.domain.enumerations;
+
+public enum MobileBrand {
+    Apple, Samsung, Huawei, Xiaomi, Nokia, Motorola, Sony, Meizu, Vivo, Honor
+}
