+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/MobileOS.java
@@ -1,0 +1,5 @@
+package com.royal.products.domain.enumerations;
+
+public enum MobileOS {
+    Android, iOS
+}
