+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/Processor.java
@@ -1,0 +1,7 @@
+package com.royal.products.domain.enumerations;
+
+public enum Processor {
+    AppleSilicon,
+    IntelCore3, IntelCore5, IntelCore7, IntelCore9,
+    AMDRyzen3, AMDRyzen5, AMDRyzen7, AMDRyzen9
+}
