+++ RoyalStore/src/main/java/com/royal/products/domain/enumerations/ProductCategory.java
@@ -1,0 +1,5 @@
+package com.royal.products.domain.enumerations;
+
+public enum ProductCategory {
+    Laptop, Smartphone
+}
