+++ RoyalStore/src/main/java/com/royal/products/repository/LaptopRepository.java
