--- RoyalStore/src/main/java/com/royal/repositories/UserRepository.java
@@ -1,12 +1,0 @@
-package com.royal.repositories;
-
-import com.royal.models.users.User;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends MongoRepository<User, String> {
-    boolean existsByEmail(String email);
-
-    Optional<User> findByEmail(String email);
-}
