+++ RoyalStore/src/main/java/com/royal/users/domain/details/AuthenticatedUserDetails.java
@@ -1,0 +1,16 @@
+package com.royal.users.details;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.ArrayList;
+
+@Getter
+@Setter
+public class AuthenticatedUserDetails {
+    private String email;
+    private String password;
+    private boolean rememberMe;
+    private byte[] profilePicture;
+    private ArrayList<String> roles;
+}
