+++ RoyalStore/src/main/java/com/royal/users/domain/details/LoginUserCredentials.java
@@ -1,0 +1,12 @@
+package com.royal.users.details;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginUserCredentials {
+    private String email;
+    private String password;
+    private boolean rememberMe;
+}
