+++ RoyalStore/src/main/java/com/royal/users/domain/details/PublicUserDetails.java
@@ -1,0 +1,17 @@
+package com.royal.users.details;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.ArrayList;
+
+@Getter
+@Setter
+public class PublicUserDetails {
+    private String id;
+    private String token;
+    private String email;
+    private String username;
+    private byte[] profilePicture;
+    private ArrayList<String> roles;
+}
