+++ RoyalStore/src/main/java/com/royal/users/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.royal.users;
+
+import com.royal.users.User;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends MongoRepository<User, String> {
+    boolean existsByEmail(String email);
+
+    Optional<User> findByEmail(String email);
+}
